Consolidate ZSTD utility functions
The zstd_compressor and zstd_decompressor implementations had a
number of duplicated utility functions between them; this consolidates
them into one file.
