+++ rosbag2_compression/src/rosbag2_compression/zstd_compressor.cpp
@@ -21,201 +21,8 @@
 
 #include "rcpputils/filesystem_helper.hpp"
 
+#include "compression_utils.hpp"
 #include "rosbag2_compression/zstd_compressor.hpp"
-
-#include "logging.hpp"
-
-namespace
-{
-// Increasing the compression level will:
-//   - Increase the time taken to compress
-//   - Decrease the size of the compressed data
-// Setting to zero uses Zstd's default value of 3.
-constexpr const int kDefaultZstdCompressionLevel = 1;
-
-// String constant used to identify ZstdCompressor.
-constexpr const char kCompressionIdentifier[] = "zstd";
-// Used as a parameter type in a function that accepts the output of ZSTD_compress.
-using ZstdCompressReturnType = decltype(ZSTD_compress(nullptr, 0, nullptr, 0, 0));
-
-/**
- * Open a file using the OS-specific C API.
- * \param uri is the path to the file.
- * \param read_mode is the read mode accepted by OS-specific fopen.
- * \return the FILE pointer or nullptr if the file was not opened.
- */
-FILE * open_file(const std::string & uri, const std::string & read_mode)
-{
-  FILE * fp{nullptr};
-#ifdef _WIN32
-  fopen_s(&fp, uri.c_str(), read_mode.c_str());
-#else
-  fp = std::fopen(uri.c_str(), read_mode.c_str());
-#endif
-  return fp;
-}
-
-/**
- * Read a file from the supplied uri into a vector.
- * \param uri is the path to the file.
- * \return the contents of the buffer as a vector.
- */
-std::vector<uint8_t> get_input_buffer(const std::string & uri)
-{
-  // Read in buffer, handling accordingly
-  const auto file_pointer = open_file(uri.c_str(), "rb");
-  if (file_pointer == nullptr) {
-    std::stringstream errmsg;
-    errmsg << "Error opening file: \"" << uri <<
-      "\" for binary reading! errno(" << errno << ")";
-
-    throw std::runtime_error{errmsg.str()};
-  }
-
-  const auto file_path = rcpputils::fs::path{uri};
-  const auto decompressed_buffer_length = file_path.exists() ? file_path.file_size() : 0u;
-
-  if (decompressed_buffer_length == 0) {
-    fclose(file_pointer);
-
-    std::stringstream errmsg;
-    errmsg << "Unable to get size of file: \"" << uri << "\"";
-
-    throw std::runtime_error{errmsg.str()};
-  }
-
-  // Allocate and read in
-  std::vector<uint8_t> decompressed_buffer(decompressed_buffer_length);
-
-  const auto read_count = fread(
-    decompressed_buffer.data(), sizeof(uint8_t), decompressed_buffer.size(), file_pointer);
-
-  if (read_count != decompressed_buffer_length) {
-    ROSBAG2_COMPRESSION_LOG_ERROR_STREAM(
-      "Bytes read (" << read_count <<
-        ") != decompressed_buffer_length (" << decompressed_buffer.size() << ")!");
-    // An error indicator is set by this, so the following check will throw
-  }
-
-  if (ferror(file_pointer)) {
-    fclose(file_pointer);
-
-    std::stringstream errmsg;
-    errmsg << "Unable to read binary data from file: \"" << uri << "\"!";
-
-    throw std::runtime_error{errmsg.str()};
-  }
-  fclose(file_pointer);
-  return decompressed_buffer;
-}
-
-/**
- * Writes the output buffer to the specified file path.
- * \param output_buffer is the data to write.
- * \param uri is the relative file path to the output storage.
- */
-void write_output_buffer(
-  const std::vector<uint8_t> & output_buffer,
-  const std::string & uri)
-{
-  if (output_buffer.empty()) {
-    std::stringstream errmsg;
-    errmsg << "Cannot write empty buffer to file: \"" << uri << "\"";
-
-    throw std::runtime_error{errmsg.str()};
-  }
-
-  const auto file_pointer = open_file(uri.c_str(), "wb");
-  const auto write_count = fwrite(
-    output_buffer.data(),
-    sizeof(uint8_t),
-    output_buffer.size(),
-    file_pointer);
-
-  if (write_count != output_buffer.size()) {
-    ROSBAG2_COMPRESSION_LOG_ERROR_STREAM(
-      "Bytes written (" << write_count <<
-        ") != output_buffer size (" << output_buffer.size() << ")!");
-    // An error indicator is set by fwrite, so the following check will throw.
-  }
-
-  if (ferror(file_pointer)) {
-    fclose(file_pointer);
-
-    std::stringstream errmsg;
-    errmsg << "Unable to write compressed data to file: \"" << uri << "\"!";
-
-    throw std::runtime_error{errmsg.str()};
-  }
-  fclose(file_pointer);
-}
-
-/**
- * Checks rcutils array resizing and throws a runtime_error if there was an error resizing.
- * \param rcutils_ret_t Result of calling rcutils
- */
-void throw_on_rcutils_resize_error(const rcutils_ret_t resize_result)
-{
-  if (resize_result == RCUTILS_RET_OK) {
-    return;
-  }
-
-  std::stringstream error;
-  error << "rcutils_uint8_array_resize error: ";
-  switch (resize_result) {
-    case RCUTILS_RET_INVALID_ARGUMENT:
-      error << "Invalid Argument";
-      break;
-    case RCUTILS_RET_BAD_ALLOC:
-      error << "Bad Alloc";
-      break;
-    case RCUTILS_RET_ERROR:
-      error << "Ret Error";
-      break;
-    default:
-      error << "Unexpected Result";
-      break;
-  }
-  throw std::runtime_error(error.str());
-}
-
-/**
- * Checks compression_result and throws a runtime_error if there was a ZSTD error.
- * \param compression_result is the return value of ZSTD_compress.
- */
-void throw_on_zstd_error(const ZstdCompressReturnType compression_result)
-{
-  if (ZSTD_isError(compression_result)) {
-    std::stringstream error;
-    error << "ZSTD compression error: " << ZSTD_getErrorName(compression_result);
-    throw std::runtime_error{error.str()};
-  }
-}
-
-/**
- * Prints compression statistics to the debug log stream.
- * The log statement is formatted in JSON.
- * Time is formatted as a decimal of seconds.
- *
- * Example:
- *   "Compression statistics: {"Time" : 1.2, "Compression Ratio" : 0.5}
- */
-void print_compression_statistics(
-  std::chrono::high_resolution_clock::time_point start,
-  std::chrono::high_resolution_clock::time_point end,
-  size_t decompressed_size, size_t compressed_size)
-{
-  const auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start);
-  const auto compression_ratio =
-    static_cast<double>(decompressed_size) / static_cast<double>(compressed_size);
-  ROSBAG2_COMPRESSION_LOG_DEBUG_STREAM(
-    "\"Compression statistics\" : {" <<
-      "\"Time\" : " << (duration.count() / 1000.0) <<
-      ", \"Compression Ratio\" : " << compression_ratio <<
-      "}"
-  );
-}
-}  // namespace
 
 namespace rosbag2_compression
 {
